[msm]: refactored package substructure
