+++ pyemma/msm/ui/__init__.py
@@ -49,10 +49,5 @@ ITS-object
    ImpliedTimescales - ITS-object
    
 """
-from .msm import MSM
-from .msm_estimated import EstimatedMSM
-from .msm_sampled import SampledMSM
-from .hmsm import HMSM
-from .hmsm_estimated import EstimatedHMSM
 from .timescales import *
 from .chapman_kolmogorov import *
